refactor(fe): purge ~968 LOC dead code (barrels, composables, UI components, infra)
- Delete 13 dead component barrel re-exports (src/components/*/index.ts)
- Delete 14 single-file backward-compat re-exports (composables/stores/schemas/types/services/plugins/views)
- Delete 6 dead model re-exports (src/models/equipment/objective/pregnancy/quest/relationship/room)
- Delete 4 dead composables (useTerminalAudio 326 LOC, useAuth, useFlickering, barrel)
- Delete 2 unused UI components (ComingSoonBadge, UDropdown)
- Delete aspirational infra + dead API chain (api.ts, types/index.ts, api/incident.ts, models/incident.ts)
- Fix 15 test imports updated for deleted paths
- Delete stale UDropdown.test.ts

Tests: 857 pass, 1 skipped, 51 files green
