--- frontend/src/components/auth/index.ts
@@ -1,3 +1,0 @@
-// Re-exports from auth module for backward compatibility
-export { default as LoginFormTerminal } from '@/modules/auth/components/LoginFormTerminal.vue'
-export { default as RegisterForm } from '@/modules/auth/components/RegisterForm.vue'
