--- frontend/src/components/common/index.ts
@@ -1,2 +1,0 @@
-// Re-exports from core for backward compatibility
-export { default as ResourceBar } from '@/core/components/common/ResourceBar.vue'
