--- frontend/src/components/dwellers/index.ts
@@ -1,7 +1,0 @@
-// Re-exports from dwellers module for backward compatibility
-export { default as DwellerBio } from '@/modules/dwellers/components/DwellerBio.vue'
-export { default as DwellerCard } from '@/modules/dwellers/components/cards/DwellerCard.vue'
-export { default as DwellerFilterPanel } from '@/modules/dwellers/components/DwellerFilterPanel.vue'
-export { default as DwellerGridItem } from '@/modules/dwellers/components/grid/DwellerGridItem.vue'
-export { default as DwellerStatusBadge } from '@/modules/dwellers/components/stats/DwellerStatusBadge.vue'
-export { default as UnassignedDwellers } from '@/modules/dwellers/components/UnassignedDwellers.vue'
