--- frontend/src/components/equipment/index.ts
@@ -1,3 +1,0 @@
-// Re-exports from combat module for backward compatibility
-export { default as OutfitCard } from '@/modules/combat/components/equipment/OutfitCard.vue'
-export { default as WeaponCard } from '@/modules/combat/components/equipment/WeaponCard.vue'
