--- frontend/src/components/exploration/index.ts
@@ -1,5 +1,0 @@
-// Re-export from exploration module for backward compatibility
-export { default as EventTimeline } from '@/modules/exploration/components/EventTimeline.vue'
-export { default as ExplorationRewardsModal } from '@/modules/exploration/components/ExplorationRewardsModal.vue'
-export { default as ExplorerCard } from '@/modules/exploration/components/ExplorerCard.vue'
-export { default as WastelandPanel } from '@/modules/exploration/components/WastelandPanel.vue'
