--- frontend/src/components/incidents/index.ts
@@ -1,3 +1,0 @@
-// Re-exports from combat module for backward compatibility
-export { default as CombatModal } from '@/modules/combat/components/incidents/CombatModal.vue'
-export { default as IncidentAlert } from '@/modules/combat/components/incidents/IncidentAlert.vue'
