--- frontend/src/components/pregnancy/index.ts
@@ -1,3 +1,0 @@
-// Re-export from social module for backward compatibility
-export { default as PregnancyCard } from '@/modules/social/components/pregnancy/PregnancyCard.vue'
-export { default as PregnancyTracker } from '@/modules/social/components/pregnancy/PregnancyTracker.vue'
