--- frontend/src/components/profile/index.ts
@@ -1,3 +1,0 @@
-// Re-exports from profile module for backward compatibility
-export { default as ProfileEditor } from '@/modules/profile/components/ProfileEditor.vue'
-export { default as ProfileStats } from '@/modules/profile/components/ProfileStats.vue'
