--- frontend/src/components/relationships/index.ts
@@ -1,4 +1,0 @@
-// Re-export from social module for backward compatibility
-export { default as ChildrenList } from '@/modules/social/components/relationships/ChildrenList.vue'
-export { default as RelationshipCard } from '@/modules/social/components/relationships/RelationshipCard.vue'
-export { default as RelationshipList } from '@/modules/social/components/relationships/RelationshipList.vue'
