--- frontend/src/components/rooms/index.ts
@@ -1,9 +1,0 @@
-// Re-exports from rooms module for backward compatibility
-export {
-  RoomGrid,
-  RoomItem,
-  RoomMenu,
-  RoomMenuItem,
-  RoomDetailModal,
-  EmptyCell,
-} from '@/modules/rooms/components'
