--- frontend/src/components/training/index.ts
@@ -1,4 +1,0 @@
-// Re-export from progression module for backward compatibility
-export { default as TrainingProgressCard } from '@/modules/progression/components/training/TrainingProgressCard.vue'
-export { default as TrainingQueuePanel } from '@/modules/progression/components/training/TrainingQueuePanel.vue'
-export { default as TrainingRoomModal } from '@/modules/progression/components/training/TrainingRoomModal.vue'
