--- frontend/src/components/vault/index.ts
@@ -1,2 +1,0 @@
-// Re-exports from vault module for backward compatibility
-export { default as HappinessDashboard } from '@/modules/vault/components/HappinessDashboard.vue'
