--- frontend/src/components/wasteland/index.ts
@@ -1,2 +1,0 @@
-// Re-export from exploration module for backward compatibility
-export { default as WastelandPanel } from '@/modules/exploration/components/WastelandPanel.vue'
