--- frontend/src/composables/useHoverPreview.ts
@@ -1,2 +1,0 @@
-// Re-export from rooms module for backward compatibility
-export { useHoverPreview } from '@/modules/rooms/composables/useHoverPreview'
