--- frontend/src/composables/useResourceWarnings.ts
@@ -1,2 +1,0 @@
-// Re-export from vault module for backward compatibility
-export * from '@/modules/vault/composables/useResourceWarnings'
