--- frontend/src/composables/useRoomInteractions.ts
@@ -1,2 +1,0 @@
-// Re-export from rooms module for backward compatibility
-export { useRoomInteractions } from '@/modules/rooms/composables/useRoomInteractions'
