+++ frontend/src/core/components/ui/index.ts
@@ -12,7 +12,6 @@ export { default as UModal } from './UModal.vue'
 export { default as UBadge } from './UBadge.vue'
 export { default as UAlert } from './UAlert.vue'
 export { default as UTooltip } from './UTooltip.vue'
-export { default as UDropdown } from './UDropdown.vue'
 export { default as USkeleton } from './USkeleton.vue'
 export { default as UTabs } from './UTabs.vue'
 export { default as USelect } from './USelect.vue'
