--- frontend/src/core/composables/index.ts
@@ -1,10 +1,0 @@
-// Core composables - shared across all modules
-export { useAuth } from './useAuth'
-export { useTokenRefresh } from './useTokenRefresh'
-export { useToast } from './useToast'
-export { useTheme } from './useTheme'
-export { useWebSocket } from './useWebSocket'
-export { useVisualEffects } from './useVisualEffects'
-export { useFlickering } from './useFlickering'
-export { useSidePanel } from './useSidePanel'
-export { usePostEventStream, useSse } from './useEventStream'
