--- frontend/src/core/composables/useAuth.ts
@@ -1,19 +1,0 @@
-import { computed } from 'vue'
-import { useAuthStore } from '@/modules/auth/stores/auth'
-
-export function useAuth() {
-  const authStore = useAuthStore()
-
-  return {
-    // State
-    isAuthenticated: computed(() => authStore.isAuthenticated),
-    user: computed(() => authStore.user),
-    token: computed(() => authStore.token),
-    // Actions
-    login: authStore.login,
-    logout: authStore.logout,
-    register: authStore.register,
-    fetchUser: authStore.fetchUser,
-    refreshAccessToken: authStore.refreshAccessToken,
-  }
-}
