--- frontend/src/core/composables/useFlickering.ts
@@ -1,10 +1,0 @@
-import { useToggle } from '@vueuse/core'
-
-export function useFlickering() {
-  const [isFlickering, toggleFlickering] = useToggle(true)
-
-  return {
-    isFlickering,
-    toggleFlickering,
-  }
-}
