--- frontend/src/models/equipment.ts
@@ -1,14 +1,0 @@
-// Re-export from combat module for backward compatibility
-export {
-  type WeaponType,
-  type WeaponSubtype,
-  type OutfitType,
-  type Gender,
-  type Rarity,
-  type ItemBase,
-  type Weapon,
-  type Outfit,
-  getRarityColor,
-  getDamageRange,
-  getOutfitBonuses,
-} from '@/modules/combat/models/equipment'
