--- frontend/src/models/incident.ts
@@ -1,8 +1,0 @@
-// Re-export from combat module for backward compatibility
-export {
-  IncidentType,
-  IncidentStatus,
-  type Incident,
-  type IncidentListResponse,
-  type IncidentResolveResponse,
-} from '@/modules/combat/models/incident'
