--- frontend/src/models/objective.ts
@@ -1,6 +1,0 @@
-// Re-export from progression module for backward compatibility
-export type {
-  Objective,
-  ObjectiveCreate,
-  ObjectiveProgressUpdate,
-} from '@/modules/progression/models/objective'
