--- frontend/src/models/pregnancy.ts
@@ -1,7 +1,0 @@
-// Re-export from social module for backward compatibility
-export type {
-  PregnancyStatus,
-  Pregnancy,
-  PregnancyProgress,
-  DeliveryResult,
-} from '@/modules/social/models/pregnancy'
