--- frontend/src/models/quest.ts
@@ -1,7 +1,0 @@
-// Re-export from progression module for backward compatibility
-export type {
-  Quest,
-  VaultQuest,
-  QuestCreate,
-  QuestUpdate,
-} from '@/modules/progression/models/quest'
