--- frontend/src/models/relationship.ts
@@ -1,8 +1,0 @@
-// Re-export from social module for backward compatibility
-export type {
-  RelationshipType,
-  Relationship,
-  CompatibilityScore,
-  RelationshipCreate,
-  RelationshipUpdate,
-} from '@/modules/social/models/relationship'
