--- frontend/src/models/room.ts
@@ -1,2 +1,0 @@
-// Re-export from rooms module for backward compatibility
-export type { Room, RoomCreate, RoomUpdate } from '@/modules/rooms/models/room'
