--- frontend/src/plugins/axios.ts
@@ -1,4 +1,0 @@
-// Re-export from core for backward compatibility
-export { default } from '@/core/plugins/axios'
-import apiClient from '@/core/plugins/axios'
-export const api = apiClient
