--- frontend/src/schemas/vault.ts
@@ -1,2 +1,0 @@
-// Re-export from vault module for backward compatibility
-export * from '@/modules/vault/schemas/vault'
