--- frontend/src/stores/pregnancy.ts
@@ -1,2 +1,0 @@
-// Re-export from social module for backward compatibility
-export { usePregnancyStore } from '@/modules/social/stores/pregnancy'
