--- frontend/src/stores/profile.ts
@@ -1,2 +1,0 @@
-// Re-export from profile module for backward compatibility
-export * from '@/modules/profile/stores/profile'
