--- frontend/src/stores/radio.ts
@@ -1,2 +1,0 @@
-// Re-export from radio module for backward compatibility
-export * from '@/modules/radio/stores/radio'
