--- frontend/src/stores/relationship.ts
@@ -1,2 +1,0 @@
-// Re-export from social module for backward compatibility
-export { useRelationshipStore } from '@/modules/social/stores/relationship'
