--- frontend/src/types/auth.ts
@@ -1,2 +1,0 @@
-// Re-export from auth module for backward compatibility
-export * from '@/modules/auth/types/auth'
