--- frontend/src/views/index.ts
@@ -1,5 +1,0 @@
-// Re-exports from modules for backward compatibility
-export { default as HomeView } from '@/modules/vault/views/HomeView.vue'
-export { default as DwellersView } from '@/modules/dwellers/views/DwellersView.vue'
-export { default as ProfileView } from '@/modules/profile/views/ProfileView.vue'
-export { default as AboutView } from '@/modules/profile/views/AboutView.vue'
