+++ frontend/tests/unit/components/auth/RegisterForm.test.ts
@@ -3,7 +3,7 @@ import { mount, flushPromises } from '@vue/test-utils'
 import { setActivePinia, createPinia } from 'pinia'
 import { createRouter, createMemoryHistory } from 'vue-router'
 import RegisterForm from '@/modules/auth/components/RegisterForm.vue'
-import { useAuthStore } from '@/stores/auth'
+import { useAuthStore } from '@/modules/auth/stores/auth'
 
 // Mock Iconify
 vi.mock('@iconify/vue', () => ({
