+++ frontend/tests/unit/components/dwellers/UnassignedDwellers.test.ts
@@ -4,7 +4,7 @@ import { setActivePinia, createPinia } from 'pinia'
 import UnassignedDwellers from '@/modules/dwellers/components/UnassignedDwellers.vue'
 import { useDwellerStore } from '@/modules/dwellers/stores/dweller'
 import { useExplorationStore } from '@/stores/exploration'
-import { useAuthStore } from '@/stores/auth'
+import { useAuthStore } from '@/modules/auth/stores/auth'
 
 // Mock Iconify
 vi.mock('@iconify/vue', () => ({
