+++ frontend/tests/unit/stores/auth.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest'
 import { setActivePinia, createPinia } from 'pinia'
-import { useAuthStore } from '@/stores/auth'
+import { useAuthStore } from '@/modules/auth/stores/auth'
 import axios from '@/core/plugins/axios'
 
 vi.mock('@/core/plugins/axios')
