+++ frontend/tests/unit/views/ProfileView.test.ts
@@ -5,8 +5,8 @@ import { createRouter, createMemoryHistory } from 'vue-router'
 import ProfileView from '@/modules/profile/views/ProfileView.vue'
 import ProfileEditor from '@/modules/profile/components/ProfileEditor.vue'
 import { LifeDeathStatistics } from '@/modules/dwellers/components/death'
-import { useProfileStore } from '@/stores/profile'
-import { useAuthStore } from '@/stores/auth'
+import { useProfileStore } from '@/modules/profile/stores/profile'
+import { useAuthStore } from '@/modules/auth/stores/auth'
 import axios from '@/core/plugins/axios'
 import type { UserProfile } from '@/models/profile'
 import type { DeathStatistics } from '@/modules/profile/stores/profile'
